requirements.txt
# Core dependencies
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Async and performance
aiofiles>=23.0.0
asyncio-throttle>=1.0.0

# Data validation and config
pydantic>=2.0.0
pydantic-settings>=2.0.0

# CLI and user experience
rich>=13.0.0
typer>=0.9.0
click>=8.0.0

# Logging and monitoring
loguru>=0.7.0
structlog>=23.0.0

# HTTP and API
httpx>=0.25.0
aiohttp>=3.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0